requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.0.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
watchdog==5.0.2
